{{ cookiecutter.project_slug }}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.project_description }}"
readme = "README.md"
requires-python = ">={{ cookiecutter.python_version }}"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.13.1",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
